chore: split mockTime into a .ts file containing the class, and a spec.ts containing the tests
