+++ src/__tests__/conductor.spec.ts
@@ -7,7 +7,7 @@ import {
 } from '../types/src'
 import { Conductor, TimelineTriggerTimeResult } from '../conductor'
 import * as _ from 'underscore'
-import { MockTime } from './mockTime.spec'
+import { MockTime } from './mockTime'
 import { ThreadedClass } from 'threadedclass'
 import { AbstractDevice } from '../devices/abstract'
 import { getMockCall } from './lib.spec'
