+++ src/devices/__tests__/casparcg.spec.ts
@@ -10,7 +10,7 @@ import {
 	Ease,
 	Direction
 } from '../../types/src'
-import { MockTime } from '../../__tests__/mockTime.spec'
+import { MockTime } from '../../__tests__/mockTime'
 import { getMockCall } from '../../__tests__/lib.spec'
 
 // usage logCalls(commandReceiver0)
