+++ src/devices/__tests__/hyperdeck.spec.ts
@@ -13,7 +13,7 @@ import {
 	MappingHyperdeckType,
 	TransportStatus
 } from '../../types/src'
-import { MockTime } from '../../__tests__/mockTime.spec'
+import { MockTime } from '../../__tests__/mockTime'
 import { ThreadedClass } from 'threadedclass'
 import { getMockCall } from '../../__tests__/lib.spec'
