+++ src/devices/__tests__/panasonicPTZ.spec.ts
@@ -10,7 +10,7 @@ import {
 	MappingPanasonicPtz,
 	MappingPanasonicPtzType
 } from '../../types/src'
-import { MockTime } from '../../__tests__/mockTime.spec'
+import { MockTime } from '../../__tests__/mockTime'
 import { ThreadedClass } from 'threadedclass'
 import { getMockCall } from '../../__tests__/lib.spec'
 const request = require('../../__mocks__/request')
