+++ src/devices/__tests__/pharos.spec.ts
@@ -7,7 +7,7 @@ import {
 	MappingPharos,
 	TimelineContentTypePharos
 } from '../../types/src'
-import { MockTime } from '../../__tests__/mockTime.spec'
+import { MockTime } from '../../__tests__/mockTime'
 import { ThreadedClass } from 'threadedclass'
 import { getMockCall } from '../../__tests__/lib.spec'
 const WebSocket = require('../../__mocks__/ws')
