+++ src/devices/__tests__/quantel.spec.ts
@@ -6,7 +6,7 @@ import {
 	MappingQuantel,
 	QuantelTransitionType
 } from '../../types/src'
-import { MockTime } from '../../__tests__/mockTime.spec'
+import { MockTime } from '../../__tests__/mockTime'
 import { ThreadedClass } from 'threadedclass'
 import { QuantelDevice, QuantelCommandType } from '../quantel'
 require('../../__tests__/lib.spec')
